requirements.txt
# --------- pytorch --------- #
# torch>=2.0.0
# lightning>=2.0.0
# torchmetrics==0.7.0

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
tensorboard=2.14.0
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
tqdm
soundfile==0.12.1
Cython==3.0.2
librosa==0.10.1
matplotlib==3.7.2
numpy==1.24.4
phonemizer==3.2.1
scipy==1.11.2
Unidecode==1.3.6
